Move language to ast
