+++ bin/lint.bash
@@ -1,5 +1,4 @@
 #! /bin/bash
 
 flake8 $(find $(dirname $0)/../nl2prog -name "*.py") \
-       $(find $(dirname $0)/../test -name "*.py") \
-       $(find $(dirname $0)/../examples -name "*.py")
+       $(find $(dirname $0)/../test -name "*.py")
