+++ configs/hearthstone/treegen_evaluate.gin
@@ -7,13 +7,13 @@ dataset/nl2prog.gin.workspace.put.value = @nl2prog.dataset.hearthstone.download(
 dataset/nl2prog.gin.workspace.get.key   = "dataset"
 
 # ActionOptions
-nl2prog.language.action.ActionOptions.retain_variadic_fields = False
-nl2prog.language.action.ActionOptions.split_non_terminal     = False
+nl2prog.ast.action.ActionOptions.retain_variadic_fields = False
+nl2prog.ast.action.ActionOptions.split_non_terminal     = False
 
 # to_action_sequence
-nl2prog.language.action.code_to_action_sequence.parse    = @nl2prog.language.python.parse
-nl2prog.language.action.code_to_action_sequence.tokenize = @nl2prog.utils.python.tokenize_token
-nl2prog.language.action.code_to_action_sequence.options  = @nl2prog.language.action.ActionOptions()
+nl2prog.ast.action.code_to_action_sequence.parse    = @nl2prog.language.python.parse
+nl2prog.ast.action.code_to_action_sequence.tokenize = @nl2prog.utils.python.tokenize_token
+nl2prog.ast.action.code_to_action_sequence.options  = @nl2prog.ast.action.ActionOptions()
 
 # Encoder
 word_encoder/nl2prog.gin.workspace.get.key             = "query_encoder"
@@ -65,7 +65,7 @@ nl2prog.utils.CommonBeamSearchSynthesizer.create.split_states            = @nl2p
 nl2prog.utils.CommonBeamSearchSynthesizer.create.model                   = @nl2prog.nn.treegen.TrainModel()
 nl2prog.utils.CommonBeamSearchSynthesizer.create.action_sequence_encoder = @action_sequence_encoder/nl2prog.gin.workspace.get()
 nl2prog.utils.CommonBeamSearchSynthesizer.create.is_subtype              = @nl2prog.language.python.is_subtype
-nl2prog.utils.CommonBeamSearchSynthesizer.create.options                 = @nl2prog.language.action.ActionOptions()
+nl2prog.utils.CommonBeamSearchSynthesizer.create.options                 = @nl2prog.ast.action.ActionOptions()
 nl2prog.utils.CommonBeamSearchSynthesizer.create.device                  = @torch.device()
 nl2prog.utils.CommonBeamSearchSynthesizer.create.max_steps               = 350
 synthesizer/nl2prog.gin.workspace.put.value                              = @nl2prog.utils.CommonBeamSearchSynthesizer.create()
