+++ configs/hearthstone/treegen_train.gin
@@ -7,13 +7,13 @@ dataset/nl2prog.gin.workspace.put.value = @nl2prog.dataset.hearthstone.download(
 dataset/nl2prog.gin.workspace.get.key   = "dataset"
 
 # ActionOptions
-nl2prog.language.action.ActionOptions.retain_variadic_fields = False
-nl2prog.language.action.ActionOptions.split_non_terminal     = False
+nl2prog.ast.action.ActionOptions.retain_variadic_fields = False
+nl2prog.ast.action.ActionOptions.split_non_terminal     = False
 
 # to_action_sequence
-nl2prog.language.action.code_to_action_sequence.parse    = @nl2prog.language.python.parse
-nl2prog.language.action.code_to_action_sequence.tokenize = @nl2prog.utils.python.tokenize_token
-nl2prog.language.action.code_to_action_sequence.options  = @nl2prog.language.action.ActionOptions()
+nl2prog.ast.action.code_to_action_sequence.parse    = @nl2prog.language.python.parse
+nl2prog.ast.action.code_to_action_sequence.tokenize = @nl2prog.utils.python.tokenize_token
+nl2prog.ast.action.code_to_action_sequence.options  = @nl2prog.ast.action.ActionOptions()
 
 # Encoder
 train_dataset/nl2prog.gin.get_key.key                  = "train"
@@ -23,7 +23,7 @@ nl2prog.gin.treegen.prepare_encoder.word_threshold     = 3
 nl2prog.gin.treegen.prepare_encoder.char_threshold     = 0
 nl2prog.gin.treegen.prepare_encoder.token_threshold    = 0
 nl2prog.gin.treegen.prepare_encoder.parse              = @nl2prog.language.python.parse
-nl2prog.gin.treegen.prepare_encoder.to_action_sequence = @nl2prog.language.action.code_to_action_sequence
+nl2prog.gin.treegen.prepare_encoder.to_action_sequence = @nl2prog.ast.action.code_to_action_sequence
 nl2prog.gin.treegen.prepare_encoder.extract_query      = @nl2prog.utils.python.tokenize_query
 nl2prog.gin.treegen.prepare_encoder.tokenize_token     = @nl2prog.utils.python.tokenize_token
 word_encoder/nl2prog.gin.workspace.get.key             = "query_encoder"
@@ -37,7 +37,7 @@ nl2prog.utils.transform.treegen.TransformQuery.word_encoder                = @wo
 nl2prog.utils.transform.treegen.TransformQuery.char_encoder                = @char_encoder/nl2prog.gin.workspace.get()
 nl2prog.utils.transform.treegen.TransformQuery.max_word_length             = 128
 nl2prog.utils.transform.TransformDataset.transform_code                    = @nl2prog.utils.transform.TransformCode()
-nl2prog.utils.transform.TransformCode.to_action_sequence                   = @nl2prog.language.action.code_to_action_sequence
+nl2prog.utils.transform.TransformCode.to_action_sequence                   = @nl2prog.ast.action.code_to_action_sequence
 nl2prog.utils.transform.TransformDataset.transform_evaluator               = @nl2prog.utils.transform.treegen.TransformEvaluator()
 nl2prog.utils.transform.treegen.TransformEvaluator.action_sequence_encoder = @action_sequence_encoder/nl2prog.gin.workspace.get()
 nl2prog.utils.transform.treegen.TransformEvaluator.max_arity               = 128
