+++ nl2prog/encoders/action_sequence_encoder.py
@@ -3,12 +3,12 @@ from torchnlp.encoders import LabelEncoder
 from typing import Any, List, Optional, Union, cast
 from dataclasses import dataclass
 
-from nl2prog.language.action import NodeType, NodeConstraint
-from nl2prog.language.action import ActionSequence, ActionOptions
-from nl2prog.language.action import ApplyRule, GenerateToken
-from nl2prog.language.action import Rule, ExpandTreeRule
-from nl2prog.language.action import CloseNode, CloseVariadicFieldRule
-from nl2prog.language.evaluator import Evaluator
+from nl2prog.ast.action import NodeType, NodeConstraint
+from nl2prog.ast.action import ActionSequence, ActionOptions
+from nl2prog.ast.action import ApplyRule, GenerateToken
+from nl2prog.ast.action import Rule, ExpandTreeRule
+from nl2prog.ast.action import CloseNode, CloseVariadicFieldRule
+from nl2prog.ast.evaluator import Evaluator
 
 
 def convert_node_type_to_key(node_type: NodeType) -> NodeType:
