+++ nl2prog/language/bash/utils.py
@@ -1,6 +1,6 @@
 from typing import Union
 
-from nl2prog.language.ast import Root
+from nl2prog.ast.ast import Root
 
 
 def is_subtype(subtype: Union[str, Root], basetype: Union[str, Root]) -> bool:
