+++ nl2prog/language/python/ast_to_python_ast.py
@@ -1,7 +1,7 @@
 import ast as python_ast  # noqa
 # from typing import List
 
-import nl2prog.language.ast as ast
+import nl2prog.ast.ast as ast
 from nl2prog.language.python import PythonAST
 from .utils import BuiltinType
