+++ nl2prog/utils/transform/treegen.py
@@ -3,7 +3,7 @@ import numpy as np
 from typing import Callable, List, Any, Optional, Tuple
 from torchnlp.encoders import LabelEncoder
 
-from nl2prog.language.evaluator import Evaluator
+from nl2prog.ast.evaluator import Evaluator
 from nl2prog.encoders import ActionSequenceEncoder
 from nl2prog.utils import Query
