+++ test/gin/test_nl2prog.py
@@ -8,7 +8,7 @@ from torch import optim
 
 from nl2prog.gin import workspace
 from nl2prog.gin.nl2prog import train, evaluate
-from nl2prog.language.ast import Leaf
+from nl2prog.ast.ast import Leaf
 from nl2prog.utils.data import ListDataset, Entry
 from nl2prog.utils.beam_search_synthesizer import Candidate
 from nl2prog.metrics import Accuracy, Bleu
