+++ test/language/python/test_ast_to_python_ast.py
@@ -1,7 +1,7 @@
 import unittest
 import ast as python_ast
 
-import nl2prog.language.ast as ast
+import nl2prog.ast.ast as ast
 
 from nl2prog.language.python.ast_to_python_ast import to_builtin_type
 from nl2prog.language.python import to_ast, to_python_ast
