+++ test/language/python/test_parse.py
@@ -2,7 +2,7 @@ import unittest
 import ast
 
 from nl2prog.language.python import to_ast, parse, unparse, ParseMode
-from nl2prog.language.ast import Node
+from nl2prog.ast.ast import Node
 
 
 class TestParse(unittest.TestCase):
