+++ test/language/python/test_python_ast_to_ast.py
@@ -1,7 +1,7 @@
 import unittest
 import ast as python_ast
 
-import nl2prog.language.ast as ast
+import nl2prog.ast.ast as ast
 
 from nl2prog.language.python.python_ast_to_ast import base_ast_type
 from nl2prog.language.python import to_ast
