+++ test/test_treegen.py
@@ -11,7 +11,7 @@ import fairseq.optim as optim
 
 from nl2prog.gin import nl2prog, treegen, optimizer, workspace
 from nl2prog.utils import Query, CommonBeamSearchSynthesizer
-from nl2prog.language.action \
+from nl2prog.ast.action \
     import ast_to_action_sequence as to_seq, ActionOptions
 from nl2prog.utils.data \
     import Collate, CollateGroundTruth, collate_none, split_none, \
