+++ test/utils/transform/test_nl2code.py
@@ -3,8 +3,8 @@ import numpy as np
 from torchnlp.encoders import LabelEncoder
 from nl2prog.utils import Query
 from nl2prog.utils.data import Entry, ListDataset, get_samples
-from nl2prog.language.ast import Node, Leaf, Field
-from nl2prog.language.action import ast_to_action_sequence
+from nl2prog.ast.ast import Node, Leaf, Field
+from nl2prog.ast.action import ast_to_action_sequence
 from nl2prog.encoders import ActionSequenceEncoder
 from nl2prog.utils.transform import TransformCode
 from nl2prog.utils.transform.nl2code import TransformQuery, TransformEvaluator
